requirements.txt
dash>=2.14.2
dash-bootstrap-components>=1.4.2
pandas>=2.2.0
plotly>=5.18.0
numpy>=1.26.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
openpyxl>=3.1.2
pyodbc>=5.0.1 